Tests pass with this version.  Adding the service/helpers file and adding Mocha as a dev dependency
