+++ package.json
@@ -14,6 +14,9 @@
   "bugs": {
     "url": "https://github.com/jonathantneal/polyfill/issues"
   },
+  "scripts": {
+	  "test": "node_modules/.bin/mocha --recursive"
+  },
   "license": "CC0 1.0 Universal License",
   "dependencies": {
     "caniuse-db": "1.0.20140710",
